UI Improvements and Zooming
Moved the scripts into a single folder for better management.
Added more UI elements to show the player which keys do what.
Added zooming in and out while in camera mode.
Updated the photo gallery to have the number of pictures at the top as well as which one the player is currently viewing.